api-gateway/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
python-dotenv
